refactor: update PopUp import structure and enhance type exports
