+++ example/src/design_system/molecules/PopUp/index.ts
@@ -1,1 +1,2 @@
 export { default as PopUp } from './PopUp';
+export type { PopUpProps } from './PopUp';
