+++ example/src/design_system/molecules/index.ts
@@ -1,6 +1,7 @@
 export { Dropdown } from './Dropdown';
 export type { DropdownOption, DropdownProps } from './Dropdown';
 export { PopUp } from './PopUp';
+export type { PopUpProps } from './PopUp';
 export { SliderWithLabel } from './SliderWithLabel';
 export type { SliderWithLabelProps } from './SliderWithLabel';
 export { Switch } from './Switch';
